+++ examples/client_slotChange.js
@@ -8,3 +8,4 @@ server.on('connect', client => {
 });
 
 server.on('listening', () => console.log('Listening on localhost:25565'));
+server.listen();