+++ examples/serverList_advanced.js
@@ -34,3 +34,4 @@ const server = new Server({
 });
 
 server.on('listening', () => console.log('Listening on localhost:25565'));
+server.listen();