+++ examples/server_simple.js
@@ -9,3 +9,4 @@ server.on('connect', client => {
 });
 
 server.on('listening', () => console.log('Listening on localhost:25565'));
+server.listen();