+++ user/chunk.js
@@ -17,7 +17,6 @@ const server = new Server({
     })
 });
 
-server.on('listening', () => console.log('Listening'))
 server.on('connect', async client => {
     for (let x = -5; x < 5; x++)
         for (let z = -5; z < 5; z++)
@@ -28,3 +27,6 @@ server.on('connect', async client => {
         name: client.username
     });
 });
+
+server.on('listening', () => console.log('Listening'));
+server.listen();