+++ user/player-2.js
@@ -20,13 +20,12 @@ const server = new Server({
 });
 
 const uuid = {
-    'boem231': '4e251f1b-5d8f-42a4-8d09-15e08ebb575a',
-    'Jeroen64': '57c28f3e-47f6-4b2d-9b32-1ce0e078f813',
-    'Notch': '069a79f4-44e9-4726-a5be-fca90e38aaf5',
-    'Dream': 'ec70bcaf-702f-4bb8-b48d-276fa52a780c'
+    boem231: '4e251f1b-5d8f-42a4-8d09-15e08ebb575a',
+    Jeroen64: '57c28f3e-47f6-4b2d-9b32-1ce0e078f813',
+    Notch: '069a79f4-44e9-4726-a5be-fca90e38aaf5',
+    Dream: 'ec70bcaf-702f-4bb8-b48d-276fa52a780c'
 }
 
-server.on('listening', () => console.log('Listening'))
 server.on('connect', async client => {
     for (let x = -5; x < 5; x++)
         for (let z = -5; z < 5; z++)
@@ -35,7 +34,7 @@ server.on('connect', async client => {
     let player = await client.entity('player', { x: 3, y: 100, z: 3, yaw: 0, pitch: 0 }, {
         name: 'Jeroen64',
         uuid: uuid.Jeroen64,
-        skinAccountUuid: uuid.Notch,
+        skinAccountUuid: uuid.Notch
     });
 
     await wait(2000);
@@ -43,10 +42,13 @@ server.on('connect', async client => {
 
     let tabItem = await client.tabItem({
         player,
-        name: 'Hoi',
+        name: 'Hoi'
         // ping: 0,
         // skinAccountUuid: uuid.Jeroen64
     });
 
-    client.on('chat', (a) => { player.name = a });
+    client.on('chat', a => { player.name = a });
 });
+
+server.on('listening', () => console.log('Listening'));
+server.listen();